moved all tests to video player
